Staging

+++ src/Screens/Admin/City.jsx
@@ -30,9 +30,13 @@ import CancelModel from "../../Components/modals/CancelModel";
 import EditCityModal from "./Modals/EditCityModal";
 import ActiveFilter from "../../assets/active_filter.png"
 import AddButton from "../../Components/common/CustomButton";
+import EditButton from "../../Components/common/buttons/EditButton";
+import DeleteButton from "../../Components/common/buttons/deleteButton";
+import useAuth from "../../context/JwtContext";
 const env_URL_SERVER = import.meta.env.VITE_ENV_URL_SERVER
 const City = () => {
     const menuRef = useRef();
+    const { user } = useAuth()
     const navigate = useNavigate();
     const {pathname} = useLocation()
     // we have the module here
@@ -73,9 +77,9 @@ const City = () => {
     };
     const fetchCountryData = async () => {
         setPageLoading(true);
-        // const data = { "user_id":  1234 };
-        const data = { "user_id": 1234, "rows": ["id", "name"], "filters": [], "sort_by": [], "order": "asc", "pg_no": 0, "pg_size": 0 };
-        const response = await APIService.getCountries(data)
+        
+        const data = {  "rows": ["id", "name"], "filters": [], "sort_by": [], "order": "asc", "pg_no": 0, "pg_size": 0 };
+        const response = await APIService.getCountries({...data,user_id : user.id})
         const result = (await response.json()).data;
         // console.log(result.data);
 
@@ -86,9 +90,8 @@ const City = () => {
     }
     const fetchStateData = async (id) => {
         console.log(id);
-        const data = { "user_id": 1234, "country_id": id };
-        // const data = {"user_id":1234,"rows":["id","state"],"filters":[],"sort_by":[],"order":"asc","pg_no":0,"pg_size":0};
-        const response = await APIService.getState(data);
+        const data = {  "country_id": id };
+        const response = await APIService.getState({...data,user_id : user.id});
         const result = (await response.json()).data;
         console.log(result)
         if (Array.isArray(result)) {
@@ -127,7 +130,7 @@ const City = () => {
         setCurrentPage((prev) => 1)
         setCurrentPages((prev) => 15)
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: tempArray,
             sort_by: [sortField],
@@ -135,7 +138,7 @@ const City = () => {
             pg_no: 1,
             pg_size: 15
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -151,7 +154,7 @@ const City = () => {
         setCurrentPages(quantity);
         setCurrentPage((prev) => 1)
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: filterState,
             sort_by: [sortField],
@@ -160,7 +163,7 @@ const City = () => {
             pg_size: Number(quantity),
             search_key: searchInput
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -223,7 +226,7 @@ const City = () => {
         setPageLoading(true);
         setCurrentPage((prev) => page)
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: filterState,
             sort_by: [sortField],
@@ -232,7 +235,7 @@ const City = () => {
             pg_size: Number(currentPages),
             search_key: searchInput
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -259,8 +262,8 @@ const City = () => {
         setCurrentPage((prev) => 1)
 
         const data = {
-            user_id: 1234,
-            rows: ["id", "city", "state", "countryid", "country"],
+            
+            rows: ["id", "city", "state", "country"],
             filters: filterState,
             sort_by: [sortField],
             order: flag ? "asc" : "desc",
@@ -268,7 +271,7 @@ const City = () => {
             pg_size: Number(currentPages),
             search_key: searchInput,
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -283,7 +286,7 @@ const City = () => {
         setSearchInput("");
         setCurrentPage((prev) => 1)
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: filterState,
             sort_by: [sortField],
@@ -292,7 +295,7 @@ const City = () => {
             pg_size: Number(currentPages),
             search_key: "",
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -312,7 +315,7 @@ const City = () => {
         setBackDropLoading(true)
         setPageLoading(true)
         const data = {
-            user_id: 1234,
+            
             rows: ["country", "state", "city", "id"],
             filters: filterState,
             sort_by: [sortField],
@@ -329,23 +332,16 @@ const City = () => {
                 "id": "ID"
             }
         };
-        const response = await APIService.getCitiesAdmin(data)
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id})
         const temp = await response.json();
         const result = temp.data;
         console.log(temp)
         if (temp.result == 'success') {
             const d = {
                 "filename": temp.filename,
-                "user_id": 1234
+                "user_id" : user.id
             }
-            fetch(`${env_URL_SERVER}download/${temp.filename}`, {
-                method: 'POST', // or the appropriate HTTP method
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify(d) // Convert the object to a JSON string
-            })
-                .then(response => {
+            APIService.download(d, temp.filename).then(response => {
                     if (!response.ok) {
                         throw new Error('Network response was not ok ' + response.statusText);
                     }
@@ -441,7 +437,7 @@ const City = () => {
         setCurrentPage((prev) => 1)
 
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: tempArray,
             sort_by: [sortField],
@@ -450,7 +446,7 @@ const City = () => {
             pg_size: Number(currentPages),
             search_key: searchInput,
         };
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -492,7 +488,7 @@ const City = () => {
         setSortField(field);
         setFlag((prev) => !prev);
         const data = {
-            user_id: 1234,
+            
             rows: ["id", "city", "state", "countryid", "country"],
             filters: filterState,
             sort_by: [field],
@@ -502,7 +498,7 @@ const City = () => {
             search_key: searchInput,
         };
 
-        const response = await APIService.getCitiesAdmin(data);
+        const response = await APIService.getCitiesAdmin({...data,user_id : user.id});
         const res = await response.json();
         const result = res.data;
         const t = res.total_count;
@@ -565,12 +561,12 @@ const City = () => {
     }
     const addCity = async () => {
         const data = {
-            "user_id": 1234,
+            
             "city": formValues.cityName,
             "state": formValues.state,
             "countryid": formValues.country
         }
-        const response = await APIService.addCities(data)
+        const response = await APIService.addCities({...data,user_id : user.id})
         const res = await response.json()
         if (res.result == 'success') {
             setShowAddConfirmation(false)
@@ -591,10 +587,9 @@ const City = () => {
     }
     const deleteCity = async (id) => {
         const data = {
-            "user_id": 1234,
             "id": id
         }
-        const response = await APIService.deleteCities(data)
+        const response = await APIService.deleteCities({...data,user_id : user.id})
         const res = await response.json()
         if (res.result === 'success') {
             // delete success
@@ -944,8 +939,17 @@ const City = () => {
                                                 <p>{item.id}</p>
                                             </div>
                                             <div className="w-1/2 0 p-4 flex justify-between items-center ml-1">
-                                                <button onClick={() => handleEdit(item)}><img className="w-5 h-5" src={Edit} alt="edit" /></button>
-                                                <button onClick={() => handleDeleteCity(item)}><img className="w-5 h-5" src={Trash} alt="trash" /></button>
+                                            <EditButton
+                                             rowData={item}
+                                             handleEdit={handleEdit}
+                                            />
+                                            <DeleteButton
+                                              
+                                              handleDelete={handleDeleteCity}
+                                              rowData={item}
+                                            />
+                                                {/* <button onClick={() => handleEdit(item)}><img className="w-5 h-5" src={Edit} alt="edit" /></button>
+                                                <button onClick={() => handleDeleteCity(item)}><img className="w-5 h-5" src={Trash} alt="trash" /></button> */}
                                             </div>
                                         </div>
                                     </div>
